requirements.txt
appdirs==1.4.3
attrs==17.4.0
click==6.7
Flask==0.12.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
packaging==16.8
passlib==1.7.1
pdfminer3k==1.3.1
pluggy==0.6.0
ply==3.10
py==1.5.2
pyparsing==2.2.0
pytest==3.4.0
six==1.10.0
SQLAlchemy==1.2.2
Werkzeug==0.12.1
WTForms==2.1
